Add tty in preparation for demultiplexing the logs
Logs have extra data prefixed when the container is not running tty. We
need to know what mode the container is in (we'll want that data later
for Attr as well) so we know whether to demultiplex.

Also assume logs will grow quickly and grow the memory to hold them
aggressively.