Make closing and reopening a file work
Previously Close had the wrong signature so we'd never close it and kept
streaming updates. Now we correctly close the input stream when the file
is closed and clear the buffer; that means closed files have their size
reset to 0 and we stop ongoing network activity.

When you open a file in vim and switch between them it closes and
reopens each file. We may want to add a delay before closing.