Fix write-to-channel error after file is closed
If the file is closed within 100ms of finishing opening it, the channel
used to confirm opened may have been closed by exiting `Stream()`
without setting `confirmed`, and a goroutine on a 100ms sleep may then
run and try to write to that channel. Always set `confirmed=true` after
closing the `confirm` channel.