Allow specific plugins to fail so wash is still usable
Don't exit if some plugins fail to load. This can happen because config
is unavailable, but other plugins should still be usable.

Exit if no plugins are loaded. Note that this rarely happens because the
Docker client will successfully initialize even if there's nothing for
it to talk to.

Fixes #23.