Move buffer maps to lowest level, use sync.Map
Maps holding buffered (or previously buffered) data are accessed
frequently, added to occasionally, and never removed. This perfectly
fits the use case for `sync.Map`. Switch to using it to simplify
locking.

Also use read/write mutexes for namespace and projects maps to ensure
we're not iterating over them when they're being updated.