Scan whole volume in one pass
Speed up volume traversal by scanning the entire volume. There's a
significant cost to starting a container in an orchestration environment
like Docker or Kubernetes, so the modest cost of scanning the entire
volume instead of a single directory seems worth it. In practice volumes
are single-purpose with a small number of files, and most operations
have to list the directory tree from the base mountpoint to the
directory being operated on anyway.