Add a lock pool to the cache
Adds a pool of locks that can be used to lock when operating on a key or
set of keys. Used to lock transactional operations that try to get a key
from cache, then if not found update the entry in the cache and return
the updated value. Avoids having a lock for every possible key by using
a hashed pool, and avoids multiple expensive requests to update the same
key.