Remove Getattr
Implementing our own `Getattr` requires also taking on generating and
tracking inodes. Not ready to do that yet to get pid logging. We'll have
to do that eventually, but for now it's breaking commands like `find`.
This and ReadDirAll will likely require updates to the upstream go fuse
library to make it work well.