Add a log file for the container
Adds <container>/log to view logs. It can be read as a file, which shows
logs up until the call to read. It can also be read as a stream, which
will show new messages (plus the last 10 lines). The Stream method can
be accessed via a new /fs/stream endpoint.

Resolves #54.