Add an expiring variable type, use for file and dir
Add a Var type that can be attached to a node to store node-specific
data with an expiration. Use it to store the reader loaded for a file
(to share attribute calculation and retrieving content) and the listing
of children of a directory for (List and Find). Restore prefetching on
directories.

This Var type may be removed in favor of a cache later to share storage
with the API.