Move /fs/list's response structure to the server; clean-up FUSE
This commit moves the LSItem struct from the client to the server. It
also cleans up some of the duplicate FUSE code, which includes
centralizing the attributes logic in a FillAttr helper. The latter's
used by the /fs/list endpoint's handler.

We also update the LSItem (ListEntry) struct to include an errors map.
This map stores any errors that were encountered when setting each of
the ListEntry object's errorable fields (currently only attributes).