Support external plugins
This commit adds support for external plugins, which can be written
in any language. They are specified in the (overridable) plugins.yaml
file as an array of hashes. Each hash consists of two keys:

* name   -- The plugin name
* script -- The plugin script that Wash will shell out to when it needs
            to invoke a particular operation on an entry.

As an example, consider the following plugins.yaml file:

- name: 'aws'
  script: 'aws.rb'

- name: 'thermostat'
  script: 'thermostat.sh'

This file specifies two external plugins: 'aws' and 'thermostat'.

The plugin script must have the following usage:

<plugin_script> <op> <path> <state> <args...>

where <op> is a Wash-specific operation such as "ls" or "open";
<path> is the entry's path relative to Wash's mountpoint; <state>
is the entry's state within the external plugin; and <args ...>
are the arguments to <op>.

<path> and <state> require a bit of explanation. If we lump them
together into a single parameter called <entry>, then the script's
usage would read as:

<plugin_script> <op> <entry> <args...>

which in OOP translates to <entry>.<op>(<args...>). Thus, <path>
and <state> represent the specific entry we're operating on within
the external plugin. <path> may be useful for very simple external
plugins, while <state> is useful for writing more complicated external
plugins. For example, <state> can represent the entry's serialization
in Ruby or Python. It should contain just enough information for the plugin
script to be able to reconstruct the entry.

For simplicity, it is recommended that plugin scripts use <path> or <state>,
but not both.

Finally, plugin scripts must implement a special <init> op, which should
output the root object's serialization.