Return the zero-value of time.Time if Unix seconds are <= 0
It turns out that the zero-value of time.Time is equivalent to
-62135596800 Unix seconds, so the existing code was not correctly
munging Unix seconds to a time.Time object. This commit fixes that
behavior by returning the zero-value of time.Time if Unix seconds
are <= 0.

Note that 0 Unix seconds corresponds to January 1st, 1970 in UTC.
It is highly unlike that external plugins will provide a time attribute
field that is set to this value, so the fix here should be safe.