Remove MemCache#Set, pin go-cache to a fork until fix is pushed upstream
16a6fd7a introduced MemCache#Set to ensure that an expired but not-yet
cleaned up entry's eviction function is invoked prior to overriding it
with a new value. Unfortunately, this code inadvertantly invoked the
eviction function for an already-cached value. In Wash's case, this
resulted in the cache returning closed journals upon a cache hit.

Unfortunately, the fix in 16a6fd7a has to land in go-cache.
https://github.com/patrickmn/go-cache/pull/75 is up to address this
issue, but until that lands, we'll have to pin go-cache to a fork.