Enable tail -f support on ExternalPluginEntry#stream
Previously, ExternalPluginEntry#stream expected the streaming process
to exit after its stdout stream was finished sending all of its data.
Thus, there was no way to poll the stdout stream for any updates, which
is required for tail -f support.

This commit wraps stdout in a reader that, when closed, closes the
stdout stream and sends a termination signal to the streaming process.
Thus, Wash can now poll the streaming process' stdout stream for updates
while also ensuring that the process is safely terminated once it is
finished polling the stream.

This commit also updates the external plugin docs on streaming + the
sshfs example to describe and illustrate the new behavior.