Temporarily disable the "Build a Docker Container" step in CI
This step keeps failing at the "download Go modules" part and is
currently blocking in-flight PRs. Since we do not yet have the
container experience fleshed out, running this step isn't really
that useful. Thus, we temporarily disable it for now and plan on
re-enabling it later once we've fleshed out the container experience.