Centralize the cache ID in CacheConfig
This way, CachedList, CachedOpen, and CachedMetadata do not require the
caller to pass-in a cache ID. Thus, plugin authors are free to use them.
CachedMetadata is especially useful for a container/ec2 instance's metadata.json
file.

This commit also paves the way for creating a more general CachedOp
method that plugin authors can use if they need more fine-grained
caching.