Remove EntryBase.Ctime; change plugin.FillAttr => plugin.Attr
This commit removes EntryBase.Ctime because it made it harder to
reason about an entry. Specifically, making the Ctime an exported field
implies that the Ctime can change, which doesn’t make sense. Creating
a SetCtime method that can only be called once is a little better, but that
requires a separate, read-only Ctime getter that’s meant to be used when
an entry with a creation time needs to override Attr(). Introducing two new
methods for the purpose of reducing only a small amount of duplication
adds an extra mental burden on plugin authors and Wash maintainers. It
also breaks the generality of the Entry interface. Specifically, an Entry does
not need to be aware of nor care about a ctime, and that fact should be reflected
in EntryBase both in its type definition and its godocs.

This commit also changes plugin.FillAttr to plugin.Attr. The latter’s easier
to read because it avoids a pointer, and it makes more sense.

Signed-off-by: Enis Inan <enis.inan@puppet.com>