Allow entry names to have '/'es by introducing cnames
Previously, plugin.NewEntry panicked when an entry name contained a '/'.
However, it turns out that there are real world use cases where entry
names can contain a '/' (e.g. Kube contexts, S3 objects). Thus, this
commit introduces the concept of canonical names, or cnames for short.
An entry's cname is its name with all '/' characters replaced by a '#'.
We chose '#' because it is on the "Characters to Avoid" list on
https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingMetadata.html.
Since S3 object keys have virtually no restrictions placed on what they
can be, it seemed reasonable to assume that if AWS was discouraging uses
of '#' characters in object keys, then it's likely that we won't have
a lot of real-world entry names containing '#' characters. However in
case that does happen, core plugin authors can specify a different '/'
replacement character using EntryBase#SetSlashReplacementChar. External
plugin authors can specify a different '/' replacement character by
specifying that character as the value of the "slash_replacement_char" key
in the entry's JSON object that is outputted by the plugin script's list
action.

CNames must be unique. CName uniqueness is checked in plugin.CachedList.
If plugin.CachedList detects duplicate CNames, it will return an error.

Finally, we extend the List endpoint to return two additional pieces of
information: the entry's API path (so that clients can use it to invoke
additional operations) and the entry's CName. We also move the munging
of the listed group's name to "." from the List endpoint to the ls
command since that's not something the API should care about.

Fixes https://github.com/puppetlabs/wash/issues/156

Signed-off-by: Enis Inan <enis.inan@puppet.com>