Delay getting region for bucket
Querying the region for every bucket when listing available buckets
results in many queries to list them. We only need the region when we're
listing the contents of the bucket. Delay getting the region name until
we need it.

The region result will be cached indefinitely. If there's an error,
delete the cache entry for the bucket to force a retry.

This saves ~4 seconds for listing an s3 account with ~180 buckets. It
now takes ~4 seconds to do the initial listing, then ~100ms for each
subsequent listing when the cache expires.

Signed-off-by: Michael Smith <michael.smith@puppet.com>