Generalize the concept of attributes and metadata
Previously, an entry’s attributes were thought to consist of
of only filesystem attributes while more specific information
about the entry (e.g. like its state, location) was delegated to its
metadata. Also, attributes and metadata were treated as separate
concepts. Specifically, entries with metadata were classified as resources.

This classification is unnecessary. An entry’s metadata completely describes
the entry. An entry’s attributes partially describe it. Thus, an entry’s attributes
must be a subset of its metadata; you should be able to determine the former
from the latter, and you should be able to refresh the former when the latter’s
refreshed. Furthermore, there is no reason why an entry’s attributes has to consist
solely of filesystem attributes. If we again think of an entry’s attributes as partially
describing it, then we see that things like an entry’s state, location, etc. are all valid
attributes. In fact, anything in the entry’s metadata can be an attribute. However,
it would be impractical (and impossible) to enumerate all the possible attributes that
any entry can have. Instead, a better approach is to “guess” some common attributes
that entries might have, and let the plugin authors fill them in. This is exactly what
plugin.EntryAttributes does. Right now, it includes only the filesystem attributes. Future
commits will add more attributes to it, starting with state.

A big benefit of plugin.EntryAttributes is that it can pave the way for a generic find.
Specifically, if we imagine find as a filter that returns only the entries satisfying a specific
set of predicates, then `wash find`’s predicates can be constructed on the attributes
specified in plugin.EntryAttributes. For example, when we add a state attribute, and if
the plugin author specifies a reasonable ctime/mtime/atime (or even a separate “startup”
time attribute), then the generic find can search for all running VMs that were started
after 1:30 PM via something like `find <vms_dir> —state=running -and —start_time > 1:30 PM`.

Sometimes, a user may want to find entries using a predicate that isn’t captured in plugin.EntryAttributes.
Since the entry’s metadata completely describes a given entry, that user can construct a predicate
on the special `meta` attribute, where `meta` is the entry’s metadata returned by Entry#Metadata.
With `meta`, the AWS reaper reduces to the following shell command:
```
wash find ec2/instances —state=running -a —meta.Tags.termination_date > [date] | wash delete
```

This commit also introduces the ability to sync an attribute to a specific key in the entry’s metadata. This way,
plugin authors do not have to manage refreshing the entry’s attributes; Wash will do it for them. However,
this does require the plugin author to ensure that the entry’s metadata includes all of its attributes. This makes
sense, since the attributes are a subset of the metadata. Furthermore, the new sync feature also lets external
plugin entries refresh their attributes, something that the previous code could not do.

Finally, this commit includes some clean-up, such as privatizing Entry#Name (so that plugin authors cannot
override it), and renaming actionOpCode to defaultOpCode.

Signed-off-by: Enis Inan <enis.inan@puppet.com>