Fix potential deadlock getting instance metadata
It's unsafe to call a cached operation from another cached operation. We
use a shared pool of locks and could end up selecting the same one for
each of those operations, causing a deadlock.

When restoring caching on ec2Instance's ListOp, we also restored it for
the MetadataOp and introduced this potential deadlock. Disable caching
Metadata as its already effectively cached with DescribeInstance. We
leave ec2InstanceMetadataJSON's call to CachedMetadata because it's up
to ec2Instance to determine which level to cache at and we want to use
whatever it's cached. Now that ec2Instance MetadataOp caching is
disabled the cache will only be used once for the DescribeInstance call
used to populate Metadata.

Signed-off-by: Michael Smith <michael.smith@puppet.com>