Fix CI errors by commenting `nolint` on grammar vars
The grammar vars are not referenced by cmdfind's parser (except for
andOp). Instead, the parser references them via the `atoms` + `binaryOps`
hashes. Hence, CI will flag them as unused variables. However, declaring
them as individual variables is still useful because we can initialize them
individually instead of through a global `init` function, and it makes
it easier for the tests to reference them when they need to. Thus, we
add a `nolint` comment above each of the variables so that CI does not
flag them.

Signed-off-by: Enis Inan <enis.inan@puppet.com>