Fix /cache, simplify get*FromRequest operations
With the change to handling full paths, we missed updating how the
`/cache` handler gets its path so it started trying to delete based on
the full system path rather than the entry ID (which is relative to
mountpoint). Fix so that it normalizes the path or errors if not a
cached path.

Also remove context as an independent argument to the `get*FromRequest`
functions. They always want the context associated with the request, so
use it directly.

Signed-off-by: Michael Smith <michael.smith@puppet.com>