Fix negative time value parsing bug in parsePrimitivePredicate
Previously, parsePrimitivePredicate failed to parse negative time values
like +{1h} because those values were marked as syntax errors by
parseNumericPredicate. {} are meant to distinguish negative numbers, so
returning syntax errors when the parser detects them is the correct
behavior.

This commit fixes parsePrimitivePredicate to try parseTimePredicate if
parseNumericPredicate returns _any_ type of error. If parseTimePredicate
successfully parses something, we return the result. If parseTimePredicate
returns a syntax error, we return that. Otherwise, if
parseNumericPredicate previously returned a syntax error, we return
that. These changes fix the bug described in the commit title.

This commit also simplifies the code by removing the grammar.go file
since with this bug-fix, the try function is no longer useful.

Signed-off-by: Enis Inan <enis.inan@puppet.com>