Return false for time-mismatches in TimePredicate
Queries like +1h reference the past, while queries like +{1h} reference
the future. Here, +1h is "more than one hour ago" while +{1h} is "more
than one hour from now". The {} determine whether the query's based on the
past or the future. Specifically, if the duration is inside the {}, then
the query's based on the future. Otherwise, the query's based on the
past.

With these semantics, one would expect a parsed TimePredicate "p" to return
false if it is given a value from a different time. For example, if the
query's based on the past, and t is a time from the future, then p(t)
should return false. Unfortunately, the previous behavior allowed for
negative 'StartTime - t' values, which meant that queries like -1h could
return true for future times 't'.

This commit fixes TimePredicate to return false for all time-mismatches
if diff < 0, where diff = 'StartTime - t' for queries based on the past (and
diff = 't - StartTime' for queries based on the future).

Signed-off-by: Enis Inan <enis.inan@puppet.com>