Normalize ExecPacketType and Exec StreamID
Exec's StreamID was an int8 type to compact representing output chunks
from plugin execution. However that complicated both representing the
stream and didn't translate to reduced bandwidth for the API because we
still represented the packet type as a string.

Unify representation of execution stream type to ExecPacketType (string)
across the plugin and api packages. This slightly increases memory used,
but is likely minimized because the struct was already padded (so int8
was probably still 8-byte aligned) and the string representation should
refer to a single instance of the const string.

Signed-off-by: Michael Smith <michael.smith@puppet.com>