Add Elevate to ExecOptions
Adds an Elevate flag to ExecOptions to request that the command be run
as a priveleged user. Use with the `fs` package to make sure we can read
system logs.

This has all sorts of ways it can fail, and starts to get into the kind
of complexity Bolt is already dealing with.

Signed-off-by: Michael Smith <michael.smith@puppet.com>