Extract server operation from setup, provide clean API
Prepare to use the wash server from other commands than just `wash
server`. Separate out the server management from subcommand processing.

Move server operation to an object with Start/Wait/Stop operations. That
allows much clearer and simpler interaction with the server.

Normalize API and FUSE server functions to return a channel that accepts
a context (and can be closed by the caller to signal stop) and a channel
that the server goroutine closes to signal completion.

Signed-off-by: Michael Smith <michael.smith@puppet.com>