Default to maxdepth 1 if the meta primary is set
The meta primary's a specialized filter. This means that for a
given entry, it is unlikely that the entry's siblings or children
will have the same meta schema. Thus, it does not make sense for
`wash find` to recurse when the user specifies the meta primary
because recursing could lead to wasteful API requests.

Thus to avoid the unnecessary recursion, this commit sets the maxdepth
option to 1 if the meta primary's specified and the flag was not
provided on the command line.

Signed-off-by: Enis Inan <enis.inan@puppet.com>