Add ability to follow Wash history
Adds `-f,--follow` option to the history command so you can watch for
new entries added to history or follow updates to a journal for an
in-progress command. This is particularly useful when a command has been
running for awhile and you want to see a more detailed look at what it's
doing.

Signed-off-by: Michael Smith <michael.smith@puppet.com>