Fix AWS region on S3 buckets
When making multiple requests to list the same S3 bucket, the first
request would get the region and update the client to use that region
but subsequent requests for the same path (but with new instances of the
s3Bucket object) would use the cached region and not update the client.
This resulted in lots of failures due to incorrect region after the
first request. Fix so we always update the client to the correct region.

Signed-off-by: Michael Smith <michael.smith@puppet.com>