Add internal.Command
internal.Command's a wrapper to exec.Cmd. It will be used by the
external plugin interface because that is where its features are needed.
internal.Command does the following:

* Properly handles context-cancellation cleanup. exec.CommandContext
sends a SIGKILL signal to the process, which does not give the process
enough time to handle its own cleanup. internal.Command sends a SIGTERM
signal first, then a SIGKILL signal after five seconds if the process is
still running. This gives external plugins time to handle their own
cleanup.

* Runs the command in its own process group. This ensures that when a
SIGTERM/SIGKILL signal is sent, the signal will propagate down to the
process' children. For external plugins, it ensures that all the process'
children are cleaned up. This issue came about from the sshfs example,
where the old code left orphaned SSH processes.

* Provides a String() method for better logging. String() includes the
PID, the PGID, and the command invocation.

Signed-off-by: Enis Inan <enis.inan@puppet.com>