Update the external plugin interface with recent changes
This commit updates the external plugin interface with the recent
attribute/metadata changes and the new internal.Command object. It
also simplifies some of the external plugin code to make it more
consistent with the documented calling convention.

The biggest change here is changing "supported_actions" => "methods".
"Methods" represents the Wash-invokable methods that are implemented by
the external plugin entry. It is more generic than "supported_actions"
because it allows external plugin entries to implement their own "metadata"
method to override EntryBase#Metadata while accounting for the fact that
"metadata" is not a Wash action.

Exec needs to be updated with the new ExecOptions. This work will be
done in a separate commit.

Signed-off-by: Enis Inan <enis.inan@puppet.com>