Reorganize README and website, provide docs page framework
Split content between the README and website more cleanly to reduce
duplication and hopefully better serve their target audiences.

The website primarily has an intro, Getting Started from releases,
Features (with links to /docs for more detail), and Contributing to
invite folks to get involved. The README has a shorter intro,
instructions for running from source, the docker-compose example, known
issues, and a roadmap.

Rather than duplicate content, pieces that seem relevant to a user that
are in the README (like the example, which is there because it requires
content from the repository) are linked from the website. Similarly
website content like docs and features that would still be useful to a
developer looking at the README are linked to avoid massive duplication.

Also start a framework for how the docs page will look.

Signed-off-by: Michael Smith <michael.smith@puppet.com>