Provide a non-interactive error when prompt for credential needed
Provides an error setting up an AWS profile that needs to prompt for MFA
credentials. Running wash via shebang or other methods that pipe stdin
should now work for operations that don't use that particular profile
rather than passing the input to the prompt.

Test Stdin in case input is piped to wash but it's still running in a
terminal.

Signed-off-by: Michael Smith <michael.smith@puppet.com>