Change NewEntry => NewEntryBase
The recent EntrySchema changes require core plugins to adopt a new
pattern. The pattern is to first declare a "base" object representing
class-level configuration like whether the entry's a singleton, its
caching configuration, etc. These "base" objects are derived from
EntryBase. Then once the class-level info is declared, the next step is
to define the constructor that takes the "base" object and fills it in
with instance-specific information like the entry's name, its
attributes, etc.

The first step towards realizing this pattern is changing NewEntry =>
NewEntryBase. Subsequent commits will update the individual plugins with
the pattern, + some of the EntrySchema stuff.

Signed-off-by: Enis Inan <enis.inan@puppet.com>