Fix bugs from previous EntrySchema PR
This commit fixes the following

* Modifies FSBase to take a name parameter representing the name of the
FS entry, and marks the entry as a singleton. Note that although the name
is an instance-level field for all of our other entries, we can be a bit
more flexible with singleton entries since there will only be one instance
of them at a time.

* Updates the EC2 Instance and Docker Container types to reflect the
FSBase changes.

* Fixes the Docker ContainersDir class' ChildSchemas method to return a
container's schema instead of its own schema.

Signed-off-by: Enis Inan <enis.inan@puppet.com>