Fix aliased commands in wash shell
Commands that are renamed to avoid colliding with builtins also have the
original names omitted when run in the wash shell. This is so their help
output is consistent within the shell. However that results in the
commands now erroring because the original command name is unknown.

Since the commands are renamed so they display correctly in help, have
the executable wrappers call to the changed name.

Signed-off-by: Michael Smith <michael.smith@puppet.com>