Dynamically reload parent data when HasSourceParent returns true
Update FUSE operations to always relist from the parent to make sure we
have fresh data for anything attached to the entry, such as attributes.

If the parent implements HasSourceAncestor and HasSourceParent returns
true, then look to its parent (recursively as needed) and requery the
hierarchy similar to how the API always gets new data. This ensures we
get fresh data when the cache has expired that originated that data,
avoiding complicated dependency tracking like the `volume` module
implemented.

Signed-off-by: Michael Smith <michael.smith@puppet.com>