Split out binary predicate ops into their own interface
This commit removes predicate.Predicate#And and predicate.Predicate#Or
due to a bug in the meta primary. Specifically, De'Morgan's law wasn't
properly being enforced. For example, De'Morgan's law states that an
expression like "! ( 5 -o 6 )" is equivalent to "! 5 -a ! 6". If we pass
a string value into this predicate, then it should return false since "!
5" and "! 6" returns false for mis-typed values. Unfortunately, since
And/Or were embedded in the genericPredicate class, this expression
returned true because "! ( 5 -o 6 )" strictly negated the inner
expression instead of splitting it up (since the inner expression was a
genericPredicate object).

Subsequent commits will add tests to the meta primary to validate that
De'Morgan's law is being properly enforced.

Note that unfortunately, we cannot create a generic "And/Or" op that
applies to all predicates because doing so loses
predicate-implementation-specific context (e.g. like the SchemaP for an
EntryPredicate and, in the meta primary schema work, the schema
predicate's key sequence information). Hopefully if/when we move to Go
2, some of Go 2's generic support can reduce the op duplication.

Signed-off-by: Enis Inan <enis.inan@puppet.com>