Update `wash find`'s main + walker with metadata schema changes
This commit consists of the following changes:

* Remove the previous "default maxdepth to 1 if meta primary's provided"
behavior. This was previously introduced to avoid unnecessary API
requests for core plugins. Now that core plugins provide metadata
schemas, this behavior is no longer needed.

* The walker will only fetch an entry's full metadata if the entry
provides a schema for it (at least one of meta-attribute or metadata).
The reason for this restriction is b/c fullmeta requests are O(N) and
hence expensive. However, this restriction does not hold for the meta
attribute b/c meta's included as part of list's result, which is orders
of magnitude cheaper than fullmeta requests.

Signed-off-by: Enis Inan <enis.inan@puppet.com>