Don't report outer expression operators as parse errors
Previously, something like "( -m -exists )" would return a parse error
because the ")" did not begin with a leading "(". This was because the
meta primary expression parser thought that ")" was part of its
expression instead of the outer expression.

This commit introduces the IncompleteOperatorError that the inner
expression parser can return. This error will pop back up to the outer
expression parser. The error's semantics are similar to the
UnknownTokenError's semantics.

Resolves https://github.com/puppetlabs/wash/issues/361

Signed-off-by: Enis Inan <enis.inan@puppet.com>