Separate EntrySchema client code from plugin author code
This commit removes EntrySchema client code from the plugin.EntrySchema
interface and moves it over to the apitypes.EntrySchema wrapper. It also
simplifies the /fs/schema?<path> endpoint's response. The new response
for a known schema is now a JSON object of <type_id> => <schema> where
<path>'s schema's TypeID is the first key (hence, a known schema's
response cannot return an empty JSON object). All other keys represent the
remaining nodes that are included in <path>'s hierarchy. This new
response is more idiomatic with graph-representation (which is what a
schema is), and it is easier to work with/parse.

The remaining changes in this commit include updating `find` and `stree`
with the refactor.

Note that the testdata/* fixtures contain examples of the new response.

Resolves https://github.com/puppetlabs/wash/issues/363

Signed-off-by: Enis Inan <enis.inan@puppet.com>