Add timeout to calls that invoke plugins, test Exec
Specific plugin invocations may never exit. That may be due to a bug in
the plugin, or because you're trying to read all of `/dev/random`. Add a
timeout and fail validation when that timeout is exceeded.

Also adds a test for Exec that runs `echo hello` on the target.

Signed-off-by: Michael Smith <michael.smith@puppet.com>