Use TypeID if provided by the entry's schema to refine search
If the entry has a schema providing a TypeID, use it to help further
distinguish between different things that behave the same. For example,
in the Docker plugin 'containers' and 'volumes' both only support List,
but they have different TypeIDs so they should be considered different
categories of thing. Include TypeID in the output so it's clear what's
being tested.

Also avoid redundant descent when a child is the same type of thing as
its parent. This prevents simple recursion like volume directories
containing more volume directories. Only added if we have a TypeID,
otherwise it would be too common to be useful.

Signed-off-by: Michael Smith <michael.smith@puppet.com>