Syncronize error handling so we get all messages
Previously some error messages might not be printed, presumably due to
buffering or race conditions in updating erred. Use a WaitGroup to
ensure we process all pending error messages before exiting. By the time
we close the channel, there should be nothing that can submit to it
because the channel was unbuffered and all tasks that might have
submitted to it are finished.

Signed-off-by: Michael Smith <michael.smith@puppet.com>