Follow symlinks with find/stat
Currently Wash doesn't know how to represent a symlink on a remote
system. We should add it, but that's more work than I want to take on
for this issue. Filed as #378.

Make it possible to represent things that are symlinks by following the
links when we explore the system and representing tham as essentially
hard links.

Signed-off-by: Michael Smith <michael.smith@puppet.com>