Update Docker and Kubernetes logs to delay reading them
Use the new behavior of setting content size if the content is cached to
ensure logs are readable by common CLI tools and skip downloading logs
when the entries are created. Logs will now show a size of 4k until
read, when they'll get an accurate size (for that moment in time).

Signed-off-by: Michael Smith <michael.smith@puppet.com>