Add website updates based on structural review
Separates goals of our website and github README. Website serves as an
introduction for those interested in using and extending Wash, README is
focused on development.

Rearranges and updates the introduction to focus a little more on what
people might accomplish with Wash.

Improve the getting started section on the website.

Resolves #381.

Signed-off-by: Michael Smith <michael.smith@puppet.com>