Use activity.Warnf for important failures
FUSE errors are often difficult to catch because commands will silently
fail and the errors will only be written to the journal. Use Warnf so
that errors that cause the command to fail are exposed in the shell.

API server errors should also be warnings, as they'll cause whatever
called them to fail.

Operations the user should know about that fail in plugin Init should
also use Warnf because otherwise they'll be hidden during startup.

Signed-off-by: Michael Smith <michael.smith@puppet.com>