Add case for the plugin.Registry in plugin.Schema
Previously, the plugin.Registry's child schemas were obtained by
invoking its ChildSchemas() method. This did not work because it would
call Entry#Schema which, for external plugins, returns nil (since it is
only there for core plugins) so stree would showcase the stub schema for
external plugins instead of their full schema. Calling plugin.Schema in the registry's
ChildSchemas is also bad because that means we'd have to swallow any external
plugin schema-generation errors.

Thus, the easiest fix is to add an extra case for the plugin.Registry in
plugin.Schema

Signed-off-by: Enis Inan <enis.inan@puppet.com>