Switch to listing GCP Storage hierarchically
List storage prefixes and objects hierarchically using the Storage API's
Query parameters. These are explicitly designed to return results in a
directory-like fashion. This works well for buckets with large numbers
of objects, but can be slower for deeply nested buckets as we perform
two queries for every directory.

The additional query - beyond listing the directory's contents - is to
get its attributes when we name the directory. Using Query means that a
prefix - directory - only has its name returned when listing things
under that prefix. Its attributes would be returned when you queried
with that prefix, but by that point we already want to have core
attributes like creation and update time. Therefore prefixes with lots
of child prefixes may be slow.

Signed-off-by: Michael Smith <michael.smith@puppet.com>