Fix panic on 'history --follow'
Fixes the panic when using history's `--follow` option. The panic was
introduced when changing the output format of whistory because we
started parsing the output rather than copying it through to stdout.
Incomplete lines at the start of output would be ignored by
`logfmt.Unmarshal`, and we would panic when time was uninitialized.

Fixes #440.

Signed-off-by: Michael Smith <michael.smith@puppet.com>