Fix concurrency issue with submitting journal invocations
Recording a method on an entry type to analytics via the Journal can
have concurrent reads and writes. We use a mutex to guard them. However
the first time we would check whether a submission had been done, we
would use a read-lock but add initialize a new entry in
`methodInvocations`, which could result in `concurrent map writes` and
`concurrent map read and map write` errors.

Separate the implementation of `methodInvoked` so it does not update
`methodInvocations`.

Fixes #496.

Signed-off-by: Michael Smith <michael.smith@puppet.com>