Fix GCS bucket object read on Linux
The fuse implementation on Linux attempts to perform block aligned
reads; this behavior conflicts with Go's `io.ReadFull`. `ReadFull`
will produce an error when the read is shorter than the buffer and does
not account for non-block aligned files.

This commit fixes the issue by performing a simple `Read` to allow for
read lengths that don't match the block size.

This fixes #532.

Signed-off-by: Adrien Thebo <adrien@lagrange-automation.io>